cleanup(various): Unused things in the clib are removed.
