+++ inc/com/centreon/concurrency/semaphore.hh
@@ -1,5 +1,5 @@
 /*
-** Copyright 2012-2013 Centreon
+** Copyright 2011-2013 Centreon
 **
 ** Licensed under the Apache License, Version 2.0 (the "License");
 ** you may not use this file except in compliance with the License.
@@ -16,16 +16,39 @@
 ** For more information : contact@centreon.com
 */
 
-#ifndef CC_CONCURRENCY_SEMAPHORE_HH
-#define CC_CONCURRENCY_SEMAPHORE_HH
+#ifndef CC_CONCURRENCY_SEMAPHORE_POSIX_HH
+#define CC_CONCURRENCY_SEMAPHORE_POSIX_HH
 
 #include <semaphore.h>
 #include "com/centreon/namespace.hh"
 
-#ifdef _WIN32
-#include "com/centreon/concurrency/semaphore_win32.hh"
-#else
-#include "com/centreon/concurrency/semaphore_posix.hh"
-#endif  // Windows or POSIX implementation.
+CC_BEGIN()
 
-#endif  // !CC_CONCURRENCY_SEMAPHORE_HH
+namespace concurrency {
+/**
+ *  @class semaphore semaphore_posix.hh "com/centreon/concurrency/semaphore.hh"
+ *  @brief Implements a semaphore.
+ *
+ *  POSIX implementation of a semaphore.
+ */
+class semaphore {
+ public:
+  semaphore(unsigned int n = 0);
+  ~semaphore() throw();
+  void acquire();
+  bool acquire(unsigned long timeout);
+  int available();
+  void release();
+  bool try_acquire();
+
+ private:
+  semaphore(semaphore const& right);
+  semaphore& operator=(semaphore const& right);
+
+  sem_t _sem;
+};
+}
+
+CC_END()
+
+#endif  // !CC_CONCURRENCY_SEMAPHORE_POSIX_HH
