+++ inc/com/centreon/concurrency/thread.hh
@@ -1,5 +1,5 @@
 /*
-** Copyright 2012-2013 Centreon
+** Copyright 2011-2013 Centreon
 **
 ** Licensed under the Apache License, Version 2.0 (the "License");
 ** you may not use this file except in compliance with the License.
@@ -16,13 +16,53 @@
 ** For more information : contact@centreon.com
 */
 
-#ifndef CC_CONCURRENCY_THREAD_HH
-#define CC_CONCURRENCY_THREAD_HH
+#ifndef CC_CONCURRENCY_THREAD_POSIX_HH
+#define CC_CONCURRENCY_THREAD_POSIX_HH
 
-#ifdef _WIN32
-#include "com/centreon/concurrency/thread_win32.hh"
-#else
-#include "com/centreon/concurrency/thread_posix.hh"
-#endif  // Windows or POSIX implementation.
+#include <mutex>
+#include <pthread.h>
+#include "com/centreon/namespace.hh"
 
-#endif  // !CC_CONCURRENCY_THREAD_HH
+CC_BEGIN()
+
+namespace concurrency {
+typedef pthread_t thread_id;
+
+/**
+ *  @class thread thread_posix.hh "com/centreon/concurrency/thread.hh"
+ *  @brief POSIX thread wrapper.
+ *
+ *  Wrap POSIX thread library (pthreads) in a nice and easy to use
+ *  class.
+ */
+class thread {
+ public:
+  thread();
+  virtual ~thread() throw();
+  void exec();
+  static thread_id get_current_id() throw();
+  static void msleep(unsigned long msecs);
+  static void nsleep(unsigned long nsecs);
+  static void sleep(unsigned long secs);
+  static void usleep(unsigned long usecs);
+  void wait();
+  bool wait(unsigned long timeout);
+  static void yield() throw();
+
+ protected:
+  virtual void _run() = 0;
+
+ private:
+  thread(thread const& right);
+  thread& operator=(thread const& right);
+  static void* _execute(void* data);
+
+  bool _initialized;
+  mutable std::mutex _mtx;
+  pthread_t _th;
+};
+}
+
+CC_END()
+
+#endif  // !CC_CONCURRENCY_THREAD_POSIX_HH
