+++ src/misc/stringifier.cc
@@ -22,11 +22,6 @@
 
 using namespace com::centreon::misc;
 
-#ifdef _WIN32
-// Standards ? Like C99 ? What for ?
-#define snprintf _snprintf
-#endif  // Win32
-
 /**
  *  Default constructor.
  *
