+++ src/timestamp.cc
@@ -17,12 +17,7 @@
 */
 
 #include <limits>
-#ifdef _WIN32
-#define NOMINMAX
-#include <windows.h>
-#else
 #include <sys/time.h>
-#endif  // Windows or POSIX.
 #include "com/centreon/timestamp.hh"
 
 using namespace com::centreon;
@@ -66,7 +61,7 @@ timestamp::~timestamp() throw() {}
 timestamp& timestamp::operator=(timestamp const& right) {
   if (this != &right)
     _internal_copy(right);
-  return (*this);
+  return *this;
 }
 
 /**
@@ -77,7 +72,7 @@ timestamp& timestamp::operator=(timestamp const& right) {
  *  @return True if equal, otherwise false.
  */
 bool timestamp::operator==(timestamp const& right) const throw() {
-  return (_secs == right._secs && _usecs == right._usecs);
+  return _secs == right._secs && _usecs == right._usecs;
 }
 
 /**
@@ -88,7 +83,7 @@ bool timestamp::operator==(timestamp const& right) const throw() {
  *  @return True if not equal, otherwise false.
  */
 bool timestamp::operator!=(timestamp const& right) const throw() {
-  return (!operator==(right));
+  return !operator==(right);
 }
 
 /**
@@ -111,7 +106,7 @@ bool timestamp::operator<(timestamp const& right) const throw() {
  *  @return True if less or equal, otherwise false.
  */
 bool timestamp::operator<=(timestamp const& right) const throw() {
-  return (operator<(right) || operator==(right));
+  return operator<(right) || operator==(right);
 }
 
 /**
@@ -122,7 +117,7 @@ bool timestamp::operator<=(timestamp const& right) const throw() {
  *  @return True if greater, otherwise false.
  */
 bool timestamp::operator>(timestamp const& right) const throw() {
-  return (!operator<=(right));
+  return !operator<=(right);
 }
 
 /**
@@ -133,7 +128,7 @@ bool timestamp::operator>(timestamp const& right) const throw() {
  *  @return True if greater or equal, otherwise false.
  */
 bool timestamp::operator>=(timestamp const& right) const throw() {
-  return (!operator<(right));
+  return !operator<(right);
 }
 
 /**
@@ -146,7 +141,7 @@ bool timestamp::operator>=(timestamp const& right) const throw() {
 timestamp timestamp::operator+(timestamp const& right) const {
   timestamp ret(*this);
   ret += right;
-  return (ret);
+  return ret;
 }
 
 /**
@@ -159,7 +154,7 @@ timestamp timestamp::operator+(timestamp const& right) const {
 timestamp timestamp::operator-(timestamp const& right) const {
   timestamp ret(*this);
   ret -= right;
-  return (ret);
+  return ret;
 }
 
 /**
@@ -172,7 +167,7 @@ timestamp timestamp::operator-(timestamp const& right) const {
 timestamp& timestamp::operator+=(timestamp const& right) {
   add_seconds(right._secs);
   add_useconds(right._usecs);
-  return (*this);
+  return *this;
 }
 
 /**
@@ -185,7 +180,7 @@ timestamp& timestamp::operator+=(timestamp const& right) {
 timestamp& timestamp::operator-=(timestamp const& right) {
   add_seconds(-right._secs);
   add_useconds(-static_cast<long>(right._usecs));
-  return (*this);
+  return *this;
 }
 
 /**
@@ -195,7 +190,6 @@ timestamp& timestamp::operator-=(timestamp const& right) {
  */
 void timestamp::add_mseconds(long msecs) {
   add_useconds(msecs * 1000);
-  return;
 }
 
 /**
@@ -205,7 +199,6 @@ void timestamp::add_mseconds(long msecs) {
  */
 void timestamp::add_seconds(time_t secs) {
   _secs += secs;
-  return;
 }
 
 /**
@@ -226,7 +219,6 @@ void timestamp::add_useconds(long usecs) {
   }
   _usecs = static_cast<unsigned int>(us);
   _transfer(&_secs, &_usecs);
-  return;
 }
 
 /**
@@ -246,7 +238,7 @@ timestamp timestamp::max_time() throw() {
   timestamp t;
   t._secs = std::numeric_limits<time_t>::max();
   t._usecs = 999999;
-  return (t);
+  return t;
 }
 
 /**
@@ -258,7 +250,7 @@ timestamp timestamp::min_time() throw() {
   timestamp t;
   t._secs = std::numeric_limits<time_t>::min();
   t._usecs = 0;
-  return (t);
+  return t;
 }
 
 /**
@@ -267,39 +259,9 @@ timestamp timestamp::min_time() throw() {
  *  @return The current timestamp.
  */
 timestamp timestamp::now() throw() {
-#ifdef _WIN32
-  // Convert Epoch to FILETIME.
-  SYSTEMTIME st_epoch;
-  st_epoch.wYear = 1970;
-  st_epoch.wMonth = 1;
-  st_epoch.wDay = 1;
-  st_epoch.wHour = 0;
-  st_epoch.wMinute = 0;
-  st_epoch.wSecond = 0;
-  st_epoch.wMilliseconds = 0;
-  FILETIME ft_epoch;
-  SystemTimeToFileTime(&st_epoch, &ft_epoch);
-
-  // Get current time as FILETIME.
-  FILETIME ft_now;
-  GetSystemTimeAsFileTime(&ft_now);
-
-  // Move times to 64-bit integers.
-  ULARGE_INTEGER large_epoch;
-  ULARGE_INTEGER large_now;
-  large_epoch.LowPart = ft_epoch.dwLowDateTime;
-  large_epoch.HighPart = ft_epoch.dwHighDateTime;
-  large_now.LowPart = ft_now.dwLowDateTime;
-  large_now.HighPart = ft_now.dwHighDateTime;
-  large_now.QuadPart -= large_epoch.QuadPart;
-  // Time is now expressed in units of 100ns since Unix Epoch.
-  return (timestamp(large_now.QuadPart / (10 * 1000000),
-                    (large_now.QuadPart % (10 * 1000000)) / 10));
-#else
   timeval tv;
   gettimeofday(&tv, NULL);
-  return (timestamp(tv.tv_sec, tv.tv_usec));
-#endif  // Windows or POSIX.
+  return timestamp(tv.tv_sec, tv.tv_usec);
 }
 
 /**
@@ -309,7 +271,6 @@ timestamp timestamp::now() throw() {
  */
 void timestamp::sub_mseconds(long msecs) {
   add_mseconds(-msecs);
-  return;
 }
 
 /**
@@ -319,7 +280,6 @@ void timestamp::sub_mseconds(long msecs) {
  */
 void timestamp::sub_seconds(time_t secs) {
   add_seconds(-secs);
-  return;
 }
 
 /**
@@ -329,7 +289,6 @@ void timestamp::sub_seconds(time_t secs) {
  */
 void timestamp::sub_useconds(long usecs) {
   add_useconds(-usecs);
-  return;
 }
 
 /**
@@ -338,7 +297,7 @@ void timestamp::sub_useconds(long usecs) {
  *  @return The time in milliseconds.
  */
 long long timestamp::to_mseconds() const throw() {
-  return (_secs * 1000ll + _usecs / 1000);
+  return _secs * 1000ll + _usecs / 1000;
 }
 
 /**
@@ -346,7 +305,7 @@ long long timestamp::to_mseconds() const throw() {
  *
  *  @return The time in seconds.
  */
-time_t timestamp::to_seconds() const throw() { return (_secs); }
+time_t timestamp::to_seconds() const throw() { return _secs; }
 
 /**
  *  Get the time in microseconds.
@@ -354,7 +313,7 @@ time_t timestamp::to_seconds() const throw() { return (_secs); }
  *  @return The time in microseconds.
  */
 long long timestamp::to_useconds() const throw() {
-  return (_secs * 1000000ll + _usecs);
+  return _secs * 1000000ll + _usecs;
 }
 
 /**************************************
@@ -371,7 +330,6 @@ long long timestamp::to_useconds() const throw() {
 void timestamp::_internal_copy(timestamp const& right) {
   _usecs = right._usecs;
   _secs = right._secs;
-  return;
 }
 
 /**
@@ -385,5 +343,4 @@ void timestamp::_transfer(time_t* secs, unsigned int* usecs) {
   // Transforms unnecessary microseconds into seconds.
   *secs += (*usecs / 1000000);
   *usecs %= 1000000;
-  return;
 }
